requirements.txt
fastapi
uvicorn
jinja2
pillow
python-multipart
